+++ test/sandbox/test/rendering_test.rb
@@ -514,6 +514,11 @@ class RenderingTest < ViewComponent::TestCase
 
     assert_includes(
       error.message,
+      "ViewComponent::Template:"
+    )
+
+    assert_includes(
+      error.message,
       "Template file and inline render method found for variant 'phone' in " \
       "VariantTemplateAndInlineVariantTemplateComponent."
     )